Crate fvm_ipld_bitfield

Modules

iter

json

Macros

bitfield

Constructs a BitField from a given list of 1s and 0s.

Structs

BitField

A bit field with buffered insertion/removal that serializes to/from RLE+. Similar to HashSet<u64>, but more memory-efficient when long runs of 1s and 0s are present.

OutOfRangeError

Enums

Error

UnvalidatedBitField

A bit field that may not yet have been validated for valid RLE+. Used to defer this validation step until when the bit field is first used, rather than at deserialization.

Traits

Validate

A trait for types that can produce a &BitField (or fail to do so). Generalizes over &BitField and &mut UnvalidatedBitField.