Struct fvm_ipld_bitfield::BitField

pub struct BitField { /* private fields */ }

A bit field with buffered insertion/removal that serializes to/from RLE+. Similar to HashSet<u64>, but more memory-efficient when long runs of 1s and 0s are present.

Implementations

impl BitField

pub fn from_bytes(bytes: &[u8]) -> Result<Self, Error>

Decodes RLE+ encoded bytes into a bit field.

pub fn to_bytes(&self) -> Vec<u8>

Turns a bit field into its RLE+ encoded form.

impl BitField

pub const fn new() -> Self

Creates an empty bit field.

pub fn from_ranges(iter: impl RangeIterator) -> Self

Creates a new bit field from a RangeIterator.

pub fn try_from_bits<I>(iter: I) -> Result<Self, OutOfRangeError>where I: IntoIterator, MaybeBitField: FromIterator<I::Item>,

Tries to create a new bitfield from a bit iterator. It fails if the resulting bitfield would contain values not in the range 0..u64::MAX (non-inclusive).

pub fn set(&mut self, bit: u64)

Adds the bit at a given index to the bit field, panicing if it’s out of range.

Panics

Panics if bit is u64::MAX.

pub fn try_set(&mut self, bit: u64) -> Result<(), OutOfRangeError>

Adds the bit at a given index to the bit field, returning an error if it’s out of range.

pub fn unset(&mut self, bit: u64)

Removes the bit at a given index from the bit field.

pub fn get(&self, index: u64) -> bool

Returns true if the bit field contains the bit at a given index.

pub fn first(&self) -> Option<u64>

Returns the index of the lowest bit present in the bit field.

pub fn last(&self) -> Option<u64>

Returns the index of the highest bit present in the bit field.

pub fn iter(&self) -> impl Iterator<Item = u64> + '_

Returns an iterator over the indices of the bit field’s set bits.

pub fn bounded_iter(&self, max: u64) -> Option<impl Iterator<Item = u64> + '_>

Returns an iterator over the indices of the bit field’s set bits if the number of set bits in the bit field does not exceed max. Returns None otherwise.

pub fn ranges(&self) -> impl RangeIterator + '_

Returns an iterator over the ranges of set bits that make up the bit field. The ranges are in ascending order, are non-empty, and don’t overlap.

pub fn is_empty(&self) -> bool

Returns true if the bit field is empty.

pub fn slice(&self, start: u64, len: u64) -> Option<Self>

Returns a slice of the bit field with the start index of set bits and number of bits to include in the slice. Returns None if the bit field contains fewer than start + len set bits.

pub fn len(&self) -> u64

Returns the number of set bits in the bit field.

pub fn cut(&self, other: &Self) -> Self

Returns a new bit field containing the bits in self that remain after “cutting” out the bits in other, and shifting remaining bits to the left if necessary. For example:

lhs:     xx-xxx--x
rhs:     -xx-x----

cut:     x  x x--x
output:  xxx--x

pub fn union<'a>(bitfields: impl IntoIterator<Item = &'a Self>) -> Self

Returns the union of the given bit fields as a new bit field.

pub fn contains_any(&self, other: &BitField) -> bool

Returns true if self overlaps with other.

pub fn contains_all(&self, other: &BitField) -> bool

Returns true if the self is a superset of other.